feat: se implementado el template para crear evento
