+++ src/components/templates/createEvent/CreateEvent.css
@@ -1,0 +1,3 @@
+.H1CreateEvent {
+    font-size: 40px;
+}
